fixup: code review updates
  - remove stagesUnlocked in favor levelEntitlements
  - removed commented out lines
  - move unlockTool from utils to reducers
