+++ src/components/Farmhand/Farmhand.js
@@ -292,7 +292,6 @@ const applyPriceEvents = (valueAdjustments, priceCrashes, priceSurges) => {
  * @property {boolean} showHomeScreen Option to show the Home Screen
  * @property {boolean} showNotifications
  * @property {farmhand.stageFocusType} stageFocus
- * @property {Object} stagesUnlocked Keys are stageFocusType, values are booleans
  * indicating if the stage has been unlocked
  * @property {Array.<farmhand.notification>} todaysNotifications
  * @property {number} todaysLosses Should always be a negative number.
@@ -421,7 +420,7 @@ export default class Farmhand extends FarmhandReducers {
   }
 
   get isForestUnlocked() {
-    return this.state.stagesUnlocked[stageFocusType.FOREST]
+    return this.levelEntitlements.stageFocusType[stageFocusType.FOREST]
   }
 
   /**
@@ -512,9 +511,6 @@ export default class Farmhand extends FarmhandReducers {
       showHomeScreen: true,
       showNotifications: true,
       stageFocus: stageFocusType.HOME,
-      stagesUnlocked: {
-        [stageFocusType.FOREST]: false,
-      },
       todaysNotifications: [],
       todaysLosses: 0,
       todaysPurchases: {},
