+++ src/game-logic/reducers/processLevelUp.js
@@ -3,8 +3,6 @@ import { levelAchieved } from '../../utils/levelAchieved'
 import {
   getRandomLevelUpReward,
   getRandomLevelUpRewardQuantity,
-  unlockStage,
-  unlockTool,
 } from '../../utils'
 import { getLevelEntitlements } from '../../utils/getLevelEntitlements'
 import { SPRINKLER_ITEM_ID } from '../../constants'
@@ -12,6 +10,7 @@ import { LEVEL_GAINED_NOTIFICATION } from '../../templates'
 
 import { addItemToInventory } from './addItemToInventory'
 import { showNotification } from './showNotification'
+import { unlockTool } from './unlockTool'
 
 /**
  * @param {farmhand.state} state
@@ -38,8 +37,6 @@ export const processLevelUp = (state, oldLevel) => {
       )
     } else if (levelObject?.unlocksTool) {
       state = unlockTool(state, levelObject.unlocksTool)
-    } else if (levelObject?.unlocksStageFocusType) {
-      state = unlockStage(state, levelObject.unlocksStageFocusType)
     }
     // This handles an edge case where the player levels up to level that
     // unlocks greater sprinkler range, but the sprinkler item is already
